vaguely making progress with querying
